More fixes and changes
Temporary pages for CV, About Me and Project. Made pictures smaller to fit in the screen without scrolling and gave the links a header. Put names for links on the cards at the top of the cards instead of bottom, changed one of the pictures to a lighter one to make it easier to read. Need to make a footer, fix phone support and make all the other pages.