New Pictures fixes additions
Made some new pages added pictures to my pages. Also changed the My Pages h2 to be beside the links so it is easier to tell what it is I am referring to. Also added an about me page where I tell people about my self, future plans and what I expected from the Wutv1 course, there is a is also a picture of me on the page. I added pictures of my friends that I took on the projects page to showcase my abilities, I also added links to further pages where people can find my game, "enhetsomvandlare" and the history of the web.